Tested Code
Still a few problems with the starting of one of the wheels from a dead
stop. Everything else works fine though. Added some code to stop
function for future reverse turning capability. May add some code for
various forward speeds but will test when full weight is applied to
wheels to see if there is a dire need for the addition.